# DidacticalEnigma.Next/Client/src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DidacticalEnigma.RestApi",
    "version": "v1"
  },
  "paths": {
...
    "/radicals/select": {
      "get": {
        "tags": [
          "Radicals"
        ],
        "operationId": "SelectRadicals",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sort",
            "type": "string"
          },
          {
            "in": "query",
            "name": "select",
            "type": "string"
          },
          {
            "in": "query",
            "name": "deselect",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/KanjiLookupResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "RadicalState": {
      "required": [
        "isAvailable",
        "isSelected",
        "radical"
      ],
      "type": "object",
      "properties": {
        "radical": {
          "type": "string"
        },
        "isAvailable": {
          "type": "boolean"
        },
        "isSelected": {
          "type": "boolean"
        }
      }
    },
    "KanjiLookupResult": {
      "required": [
        "kanji",
        "newQuery",
        "radicals"
      ],
      "type": "object",
      "properties": {
        "newQuery": {
          "type": "string"
        },
        "kanji": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "radicals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RadicalState"
          }
        }
      }
    },
...
  }
}